+++ src/Sailor.java
@@ -1,10 +1,9 @@
-
 /**
  * Enhanced interface for Sailor.
  */
 public interface Sailor extends SailorKernel {
     /**
-     * Increases the sailor's saltiness by complaining.
+     * Increases this sailor's saltiness by complaining.
      *
      * @updates this
      * @ensures saltiness of this has increased by 5
@@ -15,8 +14,7 @@ public interface Sailor extends SailorKernel {
      * Reports whether this sailor can be promoted.
      *
      * @return true iff this sailor is eligible for promotion
-     * @ensures canPromote = (experience of this >= 100 and saltiness of this <
-     *          80)
+     * @ensures canPromote = (getExperience >= 100 and getSaltiness < 80)
      */
     boolean canPromote();