+++ src/SailorKernel.java
@@ -10,7 +10,6 @@ public interface SailorKernel extends Standard<Sailor> {
       *
       * @param xp
       *             experience points to add
-      *
       * @updates this
       * @ensures experience of this has increased by xp
       */
@@ -27,12 +26,28 @@ public interface SailorKernel extends Standard<Sailor> {
      void addSaltiness(int amount);
 
      /**
-      * Advances the sailor one year 'fore the mast.
+      * Advances this sailor one year 'fore the mast.
       *
       * @updates this
-      * @ensures yearsForeTheMast of this has decreased by 1
+      * @ensures yearsForeTheMast of this is decreased by 1
       */
      void serveYear();
+
+     /**
+      * Advances this sailor to the next rank.
+      *
+      * @updates this
+      * @ensures rank of this is advanced to the next rank
+      */
+     void advanceRank();
+
+     /**
+      * Reports the experience of this sailor.
+      *
+      * @return current experience
+      * @ensures getExperience = experience of this
+      */
+     int getExperience();
 
      /**
       * Reports the current rank of this sailor.
@@ -46,7 +61,7 @@ public interface SailorKernel extends Standard<Sailor> {
       * Reports the saltiness of this sailor.
       *
       * @return current saltiness
-      * @ensures saltiness = saltiness of this
+      * @ensures getSaltiness = saltiness of this
       */
      int getSaltiness();
 
@@ -54,11 +69,7 @@ public interface SailorKernel extends Standard<Sailor> {
       * Reports the number of years remaining for this sailor.
       *
       * @return years remaining
-      * @ensures getYearsForeTheMast = yearsForeTheMast of this Reports the
-      *          number of years remaining for this sailor.
-      *
-      * @return years remaining
-      * @ensures getYearsForeTheMast = yearsForeTheMast pf this
+      * @ensures getYearsForeTheMast = yearsForeTheMast of this
       */
      int getYearsForeTheMast();
 }